renamed api folder to serverApi to avoid expo router conflict
